% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{merge.rec}
\alias{merge.rec}
\title{Recursively merge multiple data frames}
\usage{
\method{merge}{rec}(.list, ...)
}
\arguments{
\item{.list}{List of data frames to merge}

\item{...}{Additional arguments passed to base::merge()}
}
\value{
Single merged data frame
}
\description{
Helper function to recursively merge a list of data frames into a single data frame
}
\examples{
\dontrun{
df_list <- list(df1, df2, df3)
merged <- merge.rec(df_list, by = "ID", all = TRUE)
}
}
\keyword{internal}
