% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot enriched clusters on volcano plot}
\usage{
plot_clusters(.df, .pathways)
}
\arguments{
\item{.df}{Data frame with enrichment results and cluster categories}

\item{.pathways}{Data frame with pathway categories of interest}
}
\value{
ggplot object
}
\description{
Creates a volcano plot of enrichment results with highlighted clusters
}
\examples{
\dontrun{
plot <- plot_clusters(gsea_results, pathways_of_interest)
}
}
