% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_network}
\alias{plot_network}
\title{Plot network graph of enriched pathways}
\usage{
plot_network(.net, .layout, .df, .condition)
}
\arguments{
\item{.net}{igraph network object}

\item{.layout}{Network layout coordinates}

\item{.df}{Data frame with cluster information}

\item{.condition}{Name of experimental condition for labeling}
}
\value{
ggplot object
}
\description{
Creates a network visualization of clustered enriched pathways
}
\examples{
\dontrun{
network_plot <- plot_network(graph, layout, cluster_df, "HYPOXIA")
}
}
