% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{switch_alias}
\alias{switch_alias}
\title{Switch outdated gene symbol aliases to current symbols}
\usage{
switch_alias(dbconn, ID)
}
\arguments{
\item{dbconn}{Database connection with alias information}

\item{ID}{Gene symbol to check}
}
\value{
Updated gene symbol or original ID
}
\description{
Helper function to update gene symbols using alias database
}
\keyword{internal}
